fix(cloud-agent-next): isolate wrapper error parsing
